Vulns Lab/lab/requirement.txt
Flask==3.0.3
python-gnupg==0.5.2
requests==2.31.0
pillow==10.3.0
PyJWT==2.8.0